requirements.txt
click==7.1.1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
sqlparse==0.3.0
Werkzeug==1.0.1